mcp-server/requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
httpx>=0.25.0
PyGithub>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.5